requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
djangorestframework==3.3.3
gunicorn==19.4.5
psycopg2
static3==0.7.0
